more reusable and customizable useDeleteEntry
